# docs/architecture/openapi/admin-v1.yaml
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Happy Agent Platform Admin API",
    "version": "1.0.0",
    "description": "Closed control-plane contract for typed component catalogs, defaults, Agent lifecycle, evaluation, publication, sessions, runs, traces and playground."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin application"
    }
  ],
  "security": [
    {
      "agentAdminBearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "operationId": "listAgents",
        "summary": "List Agents",
        "description": "Lists Agent identities, draft versions and current published versions.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "INVALID_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Agent page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "operationId": "createAgent",
        "summary": "Create an Agent and draft",
        "description": "Creates an Agent with the minimum framework, Provider/Model and Prompt selections plus optional Tool and Skill selections.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "FRAMEWORK_NOT_SUPPORTED",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created Agent",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/admin/agents/{agentKey}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/AgentKey"
        }
      ],
      "get": {
        "tags": [
          "Agents"
        ],
        "operationId": "getAgent",
        "summary": "Read an Agent",
        "description": "Returns Agent identity and draft/published version pointers.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND"
        ],
        "responses": {
          "200": {
            "description": "Agent",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/playground/runs": {
      "post": {
        "tags": [
          "Playground"
        ],
        "operationId": "createPlaygroundRun",
        "summary": "Create a Playground run",
        "description": "Creates an isolated asynchronous Run against an explicit Agent draft revision or published version.",
        "x-asynchronous": true,
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "FRAMEWORK_NOT_SUPPORTED",
          "DEPENDENCY_NOT_CONFIGURED",
          "SAFETY_GATE_FAILED",
          "IDEMPOTENCY_CONFLICT",
          "RATE_LIMITED"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaygroundRunRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Playground Run accepted",
            "headers": {
              "Location": {
                "$ref": "#/components/headers/Location"
              },
              "Retry-After": {
                "$ref": "#/components/headers/RetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/api/v1/admin/playground/runs/{runId}/events": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RunId"
        }
      ],
      "get": {
        "tags": [
          "Playground"
        ],
        "operationId": "streamPlaygroundRunEvents",
        "summary": "Stream Playground Run events",
        "description": "Streams persisted platform events after an optional Last-Event-ID cursor.",
        "x-error-codes": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "INVALID_EVENT_CURSOR"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/LastEventId"
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-cache"
                }
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunEvent"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/api/v1/admin/playground/runs/{runId}/approvals/{approvalId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RunId"
        },
        {
          "name": "approvalId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "tags": [
          "Playground"
        ],
        "operationId": "decidePlaygroundRunApproval",
        "summary": "Approve or reject a Playground Tool call",
        "description": "Atomically records the authenticated user decision and resumes the frozen Tool call at most once.",
        "x-error-codes": [
          "VALIDATION_FAILED",
          "UNAUTHORIZED",
          "FORBIDDEN",
          "RESOURCE_NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "CONFLICT",
          "VALIDATION_FAILED"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run after applying the decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "default": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "agentAdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Platform administrator token only.",
        "x-audience": "happy-agent-admin-v1",
        "x-required-scope": "AGENT_ADMIN"
      }
    },
    "headers": {
      "ETag": {
        "description": "Strong entity tag.",
        "schema": {
          "type": "string",
          "pattern": "^\\\"[A-Za-z0-9._:-]+\\\"$"
        }
      },
      "Location": {
        "description": "Canonical URL of the created resource.",
        "schema": {
          "type": "string",
          "format": "uri-reference"
        }
      },
      "RetryAfter": {
        "description": "Suggested polling delay in seconds.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 60
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller key scoped to the administrator; a different request digest returns IDEMPOTENCY_CONFLICT.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      },
...
      "LastEventId": {
        "name": "Last-Event-ID",
        "in": "header",
        "required": false,
        "description": "Last durable event fully processed by the client.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 512
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
...
      "AgentKey": {
        "name": "agentKey",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[a-z][a-z0-9._-]{1,119}$"
        }
      },
...
      "RunId": {
        "name": "runId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "Problem": {
        "description": "Problem Details response.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Malformed request.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Administrator lacks the required scope.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource was not found.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource or idempotency state conflict.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "Unprocessable": {
        "description": "Domain validation failed.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Concurrency or rate budget exceeded.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Required dependency is not configured or available.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string",
            "maxLength": 2000
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "$ref": "#/components/schemas/ProblemCode"
          },
          "traceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "violations": {
            "type": "array",
            "maxItems": 64,
            "items": {
              "$ref": "#/components/schemas/FieldViolation"
            }
          }
        }
      },
      "CursorPageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "hasMore"
        ],
        "properties": {
          "nextCursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "ConfigEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "path",
          "value"
        ],
        "properties": {
          "path": {
            "type": "string",
            "pattern": "^[a-zA-Z][a-zA-Z0-9_.-]{0,255}$"
          },
          "value": {
            "$ref": "#/components/schemas/ConfigValue"
          }
        }
      },
...
      "VersionReference": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "componentKey",
          "version"
        ],
        "properties": {
          "componentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "version": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "ApprovalPolicy": {
        "type": "string",
        "enum": [
          "NEVER",
          "RISK_BASED",
          "ALWAYS"
        ]
      },
...
      "ToolBinding": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "toolKey",
          "contractVersion",
          "enabled"
        ],
        "properties": {
          "toolKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160
          },
          "contractVersion": {
            "type": "integer",
            "minimum": 1
          },
          "enabled": {
            "type": "boolean"
          },
          "usageGuidance": {
            "type": "string",
            "maxLength": 2000
          },
          "timeoutMs": {
            "type": "integer",
            "minimum": 100
          },
          "maxCallsPerRun": {
            "type": "integer",
            "minimum": 1
          },
          "approvalPolicy": {
            "$ref": "#/components/schemas/ApprovalPolicy"
          },
          "retryPolicy": {
            "$ref": "#/components/schemas/RetryPolicy"
          },
          "resultMode": {
            "$ref": "#/components/schemas/ResultMode"
          }
        }
      },
...
      "SkillBinding": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "skillKey",
          "version",
          "enabled"
        ],
        "properties": {
          "skillKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "enabled": {
            "type": "boolean"
          },
          "applicationConfig": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigEntry"
            }
          }
        }
      },
...
      "Agent": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "displayName",
          "applicationScope",
          "status",
          "draftRevision",
          "revision",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "applicationScope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "DISABLED",
              "RETIRED"
            ]
          },
          "draftRevision": {
            "type": "integer",
            "minimum": 1
          },
          "currentPublishedVersion": {
            "type": "integer",
            "minimum": 1
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentPage": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Agent"
            }
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        }
      },
      "CreateAgentRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "displayName",
          "applicationScope",
          "frameworkVersion",
          "providerVersion",
          "model",
          "promptVersion",
          "toolBindings",
          "skillBindings"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9._-]{1,119}$"
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "applicationScope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "frameworkVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "providerVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "model": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "promptVersion": {
            "$ref": "#/components/schemas/VersionReference"
          },
          "toolBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/ToolBinding"
            }
          },
          "skillBindings": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/SkillBinding"
            }
          }
        }
      },
...
      "RunStatus": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "WAITING_APPROVAL",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED"
        ]
      },
...
      "RunDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "runId",
          "origin",
          "agentKey",
          "status",
          "budget",
          "usage",
          "lastEventId",
          "structuredOutput",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "origin": {
            "type": "string",
            "enum": [
              "SESSION",
              "PLAYGROUND",
              "EVALUATION"
            ]
          },
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "agentVersion": {
            "type": "integer",
            "minimum": 1
          },
          "draftRevision": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "budget": {
            "$ref": "#/components/schemas/RunBudget"
          },
          "usage": {
            "$ref": "#/components/schemas/RunUsage"
          },
          "lastEventId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "structuredOutput": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigEntry"
            }
          },
          "failure": {
            "$ref": "#/components/schemas/JobFailure"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreatePlaygroundRunRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "agentKey",
          "target",
          "input",
          "businessUserId"
        ],
        "properties": {
          "agentKey": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "target": {
            "$ref": "#/components/schemas/PlaygroundTarget"
          },
          "input": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20000
          },
          "businessUserId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          }
        }
      },
      "PlaygroundTarget": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "kind",
          "revision"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED_VERSION"
            ]
          },
          "revision": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
      "AgentRunEvent": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/RunStateAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/TextDeltaAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/StructuredAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/ToolAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/HookAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/ApprovalAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/CheckpointAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorAgentRunEvent"
          },
          {
            "$ref": "#/components/schemas/CompletedAgentRunEvent"
          }
        ]
      },
...
      "ApprovalDecisionRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "decision"
        ],
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "REJECT"
            ]
          }
        }
      },
...
    }
  }
}